Added trim option.
Adding a trimming option to run_trim function and to the command line of run_trim.py. This is mostly a one-off thing to get a test set out the door while I'm at SIUE but it could be useful in the future depending on what the Astrodrizzle output end up doing.

code/run_trim.py:
 - Added the trim functionality to the run_trim function,
 - Added a command line option to toogle a trim mode. You input x_0, y_0 and boze size.