Fixed run_cosmics.py single-extension bug.
There was a bug in run_cosmics.py where the number of expected extensions was hard-coded in at 5. The 0th plus one for each chip. But some exposures only have 2 extensions (0th and PC). I used the length of the object returned by pyfits.open to determine the number of extensions.

run_cosmics.py:
 - Fixed single-extension bug.